requirements.txt
openai
flask
python-dotenv
requests
numpy
pyannote.audio
torch
torchaudio
pydub
pytest
pytest-mock